requirements.txt
SQLAlchemy~=2.0.31
sanic~=24.6.0
PyJWT~=2.8.0
asyncpg~=0.29.0
python-dotenv~=1.0.1